--- api/index.js
@@ -1,2 +1,0 @@
-export * as Halo from './halo';
-export * as Firebase from './firebase';
